requirements.txt
tkcalendar==1.6.1
pyperclip==1.8.2
Pillow==8.2.0
pycryptodome==3.10.1
